7.2.10 Updated tests. Fixed bugs using ESLint
