+++ src/components/app/app.jsx
@@ -12,7 +12,7 @@ import {getOffersByCity} from "../../reducer/data/selectors.js";
 import withActiveCard from "../../hocs/with-active-card/with-active-card.js";
 import {getCurrentCity, getOfferId, getScreenMode} from "../../reducer/application/selectors.js";
 import {Operation as DataOperation} from "../../reducer/data/data.js";
-import {Operation as UserOperation, AuthorizationStatus} from "../../reducer/user/user.js";
+import {Operation as UserOperation} from "../../reducer/user/user.js";
 import {ScreenMode} from "../../common/const.js";
 
 
@@ -54,7 +54,7 @@ class App extends PureComponent {
           <SignIn
             onSubmit={login}
           />
-        )
+        );
     }
 
     return null;
