+++ src/components/place-card-list/place-card-list.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import PropTypes from "prop-types";
 import PlaceCard from "../place-card/place-card.jsx";
 import offerPropTypes from "../../prop-types/offer-prop-types.js";
-import {ClassNames} from "../../common/const.js"
+import {ClassNames} from "../../common/const.js";
 
 
 const PlaceCardList = (props) => {
