+++ src/components/place-details/place-details.jsx
@@ -35,6 +35,9 @@ const PlaceDetails = (props) => {
 
   return (
     <div className="page">
+
+      <Header />
+
       <main className="page__main page__main--property">
         <section className="property">
           <div className="property__gallery-container container">
