+++ src/reducer/application/selectors.js
@@ -3,7 +3,7 @@ import NameSpace from "../name-space.js";
 
 export const getScreenMode = (state) => {
   return state[NameSpace.APPLICATION].screenMode;
-}
+};
 
 export const getCurrentCity = (state) => {
   return state[NameSpace.APPLICATION].currentCity;
