+++ src/reducer/user/selectors.js
@@ -7,4 +7,4 @@ export const getAuthorizationStatus = (state) => {
 
 export const getUserData = (state) => {
   return state[NameSpace.USER].user;
-}
+};
